requirements.txt
# Core dependencies
torch>=1.9.0
transformers>=4.20.0
datasets>=2.0.0
accelerate>=0.20.0

# API clients
groq>=0.4.0
openai>=1.0.0

# Data handling
pandas>=1.3.0
numpy>=1.21.0

# Configuration
pyyaml>=6.0
python-dotenv>=0.19.0

# Evaluation and metrics
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Utilities
tqdm>=4.64.0
colorama>=0.4.4
rich>=12.0.0

# Optional: For better logging
loguru>=0.6.0

# FAISS Memory Module
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4  # Use faiss-gpu if you have CUDA